First commit
This is a folder with necessary files for compile an example source code on Linux machine with GNU Arm Embedded Tooldchain.
These files were obtained from various sources, and some were created from scratch. So they do not have any type of optimization, and their function is mainly, for example, to take the first steps in the compilation of source code for ARM FM4 on a Linux machine.